Revert "bpf: nat: move snat_v{4,6}_needed into lib/nat.h"
Moving snat_v{4,6}_needed (now snat_v4_prepare_state and snat_v6_needed)
into lib/nat.h caused some tests to fail, probably due to some #ifdefs
not defined correctly.

Revert this change while we find a proper solution.
